@@ -1,11 +1,12 @@
 module github.com/syncthing/syncthing
 
-go 1.18
+go 1.19
 
 require (
-	github.com/AudriusButkevicius/pfilter v0.0.10
+	github.com/AudriusButkevicius/pfilter v0.0.11
 	github.com/AudriusButkevicius/recli v0.0.6
 	github.com/alecthomas/kong v0.7.1
+	github.com/calmh/incontainer v0.0.0-20221224152218-b3e71b103d7a
 	github.com/calmh/xdr v1.1.0
 	github.com/ccding/go-stun v0.1.4
 	github.com/certifi/gocertifi v0.0.0-20210507211836-431795d63e8d // indirect
